+++ src/textual/demo.py
@@ -14,7 +14,7 @@ from rich.text import Text
 
 from textual.app import App, ComposeResult
 from textual.binding import Binding
-from textual.containers import Container, Horizontal
+from textual.containers import Container, Horizontal, ScrollableContainer
 from textual.reactive import reactive
 from textual.widgets import (
     Button,
@@ -189,7 +189,7 @@ JSON_EXAMPLE = """{
 """
 
 
-class Body(Container):
+class Body(ScrollableContainer):
     pass